+++ src/components/PostDetail/index.ts
@@ -1,5 +1,5 @@
-export { default as PostCategory } from './PostCategory';
 export { default as PostContents } from './PostContents';
 export { default as PostInfo } from './PostInfo';
-export { default as PostProfile } from './PostProfile';
 export { default as TravelInfo } from './TravelInfo';
+export { default as TravelInfoItem } from './TravelInfoItem';
+export { default as TravelogueInfo } from './TravelogueInfo';
